workflowui/backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.4.2
SQLAlchemy==2.0.21
python-dateutil==2.8.2
PyJWT==2.10.1
gunicorn==21.2.0
